+++ Transformation/Interfaces/IReader.cs
@@ -13,6 +13,6 @@ namespace TransformationCore.Interfaces
     public interface IReader
     {
         void Initialise(string fileName,XElement config, ILogger logger);
-        void Load(BlockingCollection<Dictionary<string, object>> inputQueue, ref int errorCount, CancellationToken ct, ILogger logger, Action<bool, bool, int, string> rowLogAction);
+        void Load(BlockingCollection<Dictionary<string, object>> inputQueue, ref int errorCount, CancellationToken ct, ILogger logger, Action<bool, bool, long, string> rowLogAction);
     }
 }
